requirements.txt
flask>=3.0.0
Pillow>=10.0.0
